backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.5.0
pymysql==1.1.1
django-cors-headers
cryptography==45.0.4
drf-spectacular==0.28.0
Pillow==11.2.1